moved the encoder to a separate file
